use unsigned instead of unsigned long long

+++ jitlinkc++/rtdyld.cpp
@@ -30,7 +30,7 @@ extern "C"
   };
 
   typedef void *(*getfn_ptr)(void *, const char *, size_t);
-  typedef void (*offset_ptr)(void *, const char *, size_t, unsigned long long);
+  typedef void (*offset_ptr)(void *, const char *, size_t, unsigned);
 
   typedef AllocBlockSliceRTDYLD (*allocate_t)(void *, AllocRequestRTDYLD, SectionNameRTDYLD);
 
@@ -113,7 +113,7 @@ public:
       void *fnptr = rt->getfnPtr(rt->state, data, len);
 
       results[symbol] = JITEvaluatedSymbol(
-          JITTargetAddress((unsigned long long)fnptr),
+          JITTargetAddress((unsigned)fnptr),
           JITSymbolFlags::Exported);
     }
